pyproject.toml
[project]
name = "trex-framework"
version = "0.1.0"
description = "TREX (Treacle Relational Endpoint eXchange) Framework is a scalable system designed to seamlessly connect distributed agents and endpoints to a central server. It enables real-time communication, customizable feature integration, and efficient management of telemetry data, making it ideal for dynamic, high-performance distributed systems."
authors = [
    {name = "Adityavarman Manjunath"}
]
license = {text = "Proprietary License"}
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "rich (>=14.0.0,<15.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "aio-pika (>=9.5.5,<10.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "commentjson (>=0.9.0,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
